Change distance type ulong => ushort
The distance will never exceed ValidatorCycle - 1,
which is 1007 and can fit within 2 bytes (ushort).

+++ source/agora/consensus/ValidatorSet.d
@@ -367,7 +367,7 @@ public class ValidatorSet
 
     ***************************************************************************/
 
-    public bool hasPreimage (const ref Hash enroll_key, ulong distance) @safe
+    public bool hasPreimage (const ref Hash enroll_key, ushort distance) @safe
         nothrow
     {
         bool result = false;
